commit tree 7d82fd9f3b4128c0638cefe87e302b6e232a27d1

+++ app/controllers/users_controller.rb
@@ -10,13 +10,15 @@ class UsersController < ApplicationController
     
     post '/signup' do
       #redirect_if_logged_in
-      user = User.new(params[:user])
-      if user.save
-        session[:user_id] = user.id
-        redirect to "/tweets"
-      else
-        redirect to "/signup"
+      if params[:username] != "" && params[:email] != "" && params[:password] != ""
+        user = User.new(params)
+        if user.save
+          session[:user_id] = user.id
+          redirect to "/tweets"
+          return
+        end
       end
+      redirect to "/signup"
     end
     
     get '/login' do
@@ -31,24 +33,25 @@ class UsersController < ApplicationController
     post '/login' do
       #redirect_if_logged_in
 
-      user = User.find_by(:username => params[:user][:username])
+      user = User.find_by(:username => params[:username])
 
-      if user && user.authenticate(params[:user][:password])
+      if user && user.authenticate(params[:password])
         session[:user_id] = user.id 
         redirect to "/tweets"
       else
-        redirect to "/login"
+        redirect to "/signup"
       end
     end
 
     get '/logout' do
       #binding.pry
-      redirect_if_not_logged_in
-      erb :'/users/logout'
-    end
-    post '/logout' do
-      session.delete(:user_id)
-      redirect to "/login"
+      #redirect_if_not_logged_in
+      if logged_in?
+        session.delete(:user_id)
+        redirect to "/login"
+      else
+        redirect to '/'
+      end
     end
 
     get '/users/:slug' do
